+++ hdf/test/conv.c
@@ -78,7 +78,7 @@ static const char *test_name[] =
 /* for those machines with imprecise IEEE<-> conversions, this should be */
 /* close enough */
 #define EPS64          ((double)1.0E-14)
-#define EPS32          ((float)1.0E-7)
+#define EPS32          (1.0E-7f)
 void
 test_conv(void)
 {
