+++ hdf/util/hdf8to24.c
@@ -195,7 +195,7 @@ main(int argc, char *argv[])
     intn        do_jpeg = FALSE;    /* flag to indicate JPEG compression */
     intn        jpeg_qual = 75; /* JPEG quality factor */
     intn        do_scale = FALSE;   /* flag to indicate whether to scale images */
-    float     img_scale = (float) 1.0;  /* scaling factor */
+    float       img_scale = 1.0f;  /* scaling factor */
     int32       xdim, ydim;     /* dimensions of the image to convert */
     intn        ispal;          /* whether there's a palette with the image */
     uint8      *img_buf;        /* buffer to store the image in */
@@ -217,7 +217,7 @@ main(int argc, char *argv[])
                 switch (argv[file][1])
                   {
                       case 's':
-                          if ((img_scale = (float) atof(&argv[file][2])) <= (float)0.0)
+                          if ((img_scale = (float) atof(&argv[file][2])) <= 0.0f)
                             {   /* check for valid scale */
                                 printf("Bad scale, must be greater than 0\n");
                                 return (1);
